+++ src/App.vue
@@ -19,3 +19,7 @@
     },
   };
 </script>
+
+<style lang="scss" src="./global-style.scss">
+
+</style>