+++ src/components/WorkProgress.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="row text-center">
-    <p>Work in progress...</p>
-  </div>
-</template>
-
-<style lang="scss" scoped>
-
-</style>