+++ src/components/header/Header.vue
@@ -3,129 +3,92 @@
     <nav class="navbar" role="navigation" aria-label="main navigation">
       <div class="navbar-brand">
         <router-link :to="{ path: '/', name: 'homepage-component' }" class="navbar-item">
-          <h1 class="title is-3">Vuemmerce</h1>
+          <h1 class="title is-3 is-flex-mobile"></h1>
         </router-link>
 
-        <a role="button" class="navbar-burger burger" aria-label="menu" aria-expanded="false" data-target="navbarBasicExample">
+        <a role="button" class="navbar-burger burger" @click="isMenuOpen = !isMenuOpen" aria-label="menu" aria-expanded="false">
           <span aria-hidden="true"></span>
           <span aria-hidden="true"></span>
           <span aria-hidden="true"></span>
         </a>
       </div>
 
-      <div class="navbar-start search-wrapper">
-        <div class="control has-icons-left has-icons-right">
-          <input class="input" type="text" placeholder="Search...">
-          <span class="icon is-small is-left">
-            <i class="fas fa-search"></i>
-          </span>
+      <div class="navbar-menu is-active">
+        <div class="navbar-start">
+          <div class="navbar-item field">
+            <p class="control has-icons-left">
+              <input class="input is-rounded" type="text" placeholder="Search...">
+              <span class="icon is-small is-left">
+                <i class="fas fa-search"></i>
+              </span>
+            </p>
+          </div>
+        </div>
+        
+        <div class="navbar-end">
+          <div class="navbar-item shopping-cart" @click="showCheckoutModal">
+            <span class="icon">
+              <i class="fa fa-shopping-cart"></i>
+            </span>
+            <span :class="[numProductsAdded > 0 ? 'tag is-info' : '']">{{ numProductsAdded }}</span>
+          </div>
         </div>
       </div>
 
-      <div class="navbar-end">
-        <div class="navbar-item">
-          <div class="shopping-cart" @click="showCheckoutModal">
-            <i class="fa fa-shopping-cart"></i> {{ numProductsAdded }}
-          </div>
-          <div class="buttons">
-            <a v-if="!isUserLoggedIn" class="button is-primary" @click="showRegistrationModal">
-              <strong>Sign up</strong>
-            </a>
-            <a v-if="!isUserLoggedIn" class="button is-light" @click="showLoginModal">
-              Log in
-            </a>
-            <div v-if="isUserLoggedIn" class="navbar-item has-dropdown is-hoverable">
-              <a class="navbar-link has-text-white">
-                Welcome User
-              </a>
+      <!-- For mobile and tablet -->
+      <div v-show="isMenuOpen" class="navbar-end">
+        <menu-component></menu-component>
+      </div>
 
-              <div class="navbar-dropdown">
-                <a class="navbar-item">
-                  My account
-                </a>
-                <a class="navbar-item">
-                  Wishlist
-                </a>
-                <hr class="navbar-divider">
-                <a class="navbar-item">
-                  Log out
-                </a>
-              </div>
-            </div>
-          </div>
-        </div>
+      <!-- For desktop -->
+      <div class="navbar-end is-hidden-touch">
+        <menu-component></menu-component>
       </div>
     </nav>
-    <login-modal-component :isLoginActive.sync="isLoginActive"></login-modal-component>
-    <registration-modal-component :isRegistrationActive.sync="isRegistrationActive"></registration-modal-component>
     <checkout-modal-component :isCheckoutActive.sync="isCheckoutActive"></checkout-modal-component>
   </div>
 </template>
 
 <script>
-  import LoginModal from '../modal/Login';
-  import RegistrationModal from '../modal/Registration';
   import CheckoutModal from '../modal/Checkout';
+  import Menu from '../menu/Menu';
 
   export default {
     name: 'header-component',
 
     data () {
       return {
-        isLoginActive: false,
-        isRegistrationActive: false,
-        isCheckoutActive: false
+        isCheckoutActive: false,
+        isMenuOpen: false
       }
     },
 
     computed: {
       numProductsAdded () {
         return this.$store.getters.productsAdded.length;
-      },
-      isUserLoggedIn () {
-        return this.$store.getters.isUserLoggedIn;
       }
     },
 
     components: {
-      'login-modal-component': LoginModal,
-      'registration-modal-component': RegistrationModal,
-      'checkout-modal-component': CheckoutModal
+     
+      'checkout-modal-component': CheckoutModal,
+      'menu-component': Menu
     },
 
     methods: {
-      showLoginModal () {
-        this.isLoginActive = true;
-      },
-      showRegistrationModal () {
-        this.isRegistrationActive = true;
-      },
       showCheckoutModal () {
-        this.isCheckoutActive = true;
-      },
+          this.isCheckoutActive = true;
+      }
     }
   };
 </script>
 
 <style lang="scss" scoped>
-  .search-wrapper {
-    margin: 8px auto;
-    width: 50%;
-
-    .control { width: 100%; }
+  .title {
+    background: url('../../../public/assets/vuemmerce-logo.png') no-repeat;
+    background-position: 50% 50%;
+    background-size: 165px;
+    width: 175px;
+    height: 35px;
   }
-  .title { color: white; }
-  .navbar {
-    background: rgb(46, 149, 212);
-  }
-  .shopping-cart {
-    position: relative;
-    right: 50px;
-    color: white;
-    cursor: pointer;
-  }
-  .has-text-white:hover,
-  .navbar-item.has-dropdown:hover .navbar-link {
-    color: black!important;
-  } 
 </style>