+++ src/components/product_detail/ProductDetail.vue
@@ -1,36 +1,45 @@
 <template>
   <div class="section">
-    <div class="columns">
-      <div class="column is-one-third">
-        <figure class="levle-item image">
-          <img src="https://placeimg.com/400/200/any">
+    <div class="card is-clearfix columns">
+        <figure class="card-image is-480x480 column is-one-thirds">
+          <img src="https://bulma.io/images/placeholders/480x480.png">
         </figure>
-      </div>
-      <div class="column is-two-third">
-        <h2 class="title is-4">{{ title }}</h2>
-        <span class="title is-3"><strong>{{ price }}&euro;</strong></span>
-        <div class="is-pulled-right">
-          <button class="button is-primary" v-if="!isAddedBtn" @click="addToCart(id)">{{ addToCartLabel }}</button>
-          <button class="button is-text" v-if="isAddedBtn" @click="removeFromCart(id)">{{ removeFromCartLabel }}</button>
-        </div>
-      </div>
-    </div>
-    <div class="box__details row">
-      <div class="box__details__description">
-        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut
-          labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi
-          ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum
-          dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident,
-          sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
+        <div class="card-content column is-two-thirds">
+          <h2 class="title is-4">{{ title }}</h2>
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
+            Ut enim ad minim veniam, quis nostrud</p>
+          <div>
+            <i v-if="rating === 1" class="fa fa-star"></i>
+            <i v-if="rating === 2" class="fa fa-star"></i>
+            <i v-if="rating === 2" class="fa fa-star"></i>
+            <i v-if="rating === 3" class="fa fa-star"></i>
+            <i v-if="rating === 3" class="fa fa-star"></i>
+            <i v-if="rating === 3" class="fa fa-star"></i>
+            <i v-if="rating === 4" class="fa fa-star"></i>
+            <i v-if="rating === 4" class="fa fa-star"></i>
+            <i v-if="rating === 4" class="fa fa-star"></i>
+            <i v-if="rating === 4" class="fa fa-star"></i>
+            <i v-if="rating === 5" class="fa fa-star"></i>
+            <i v-if="rating === 5" class="fa fa-star"></i>
+            <i v-if="rating === 5" class="fa fa-star"></i>
+            <i v-if="rating === 5" class="fa fa-star"></i>
+            <i v-if="rating === 5" class="fa fa-star"></i>
+            <p>{{ reviews > 0 ? `${reviews} Reviews` : 'No reviews' }}</p>
+          </div>
+          <span class="title is-3"><strong>{{ price }}&euro;</strong></span>
+          <div class="is-pulled-right">
+            <button class="button is-primary" v-if="!isAddedBtn" @click="addToCart(id)">{{ addToCartLabel }}</button>
+            <button class="button is-text" v-if="isAddedBtn" @click="removeFromCart(id)">{{ removeFromCartLabel }}</button>
+          </div>
       </div>
     </div>
   </div>
-</template>l
+</template>
 
 <script>
 export default {
   name: 'product-detail-component',
-  props: ['id', 'title', 'price'],
+  props: ['id', 'title', 'price', 'rating', 'reviews'],
   
   data () {
     return {
@@ -67,5 +76,3 @@ export default {
   }
 };
 </script>
-
-<style lang="scss" scoped></style>