+++ src/components/user/Wishlist.vue
@@ -1,0 +1,22 @@
+<template>
+    <div>
+        <p>Wishlist</p>
+    </div>
+</template>
+
+<script>
+export default {
+    name: 'wishlist-component',
+    data () {
+        return {
+
+        }
+    }
+}
+</script>
+
+<style lang="scss" scoped>
+
+</style>
+
+