+++ changelog.md
@@ -1,5 +1,15 @@
 # Vuemmerce Changelog
 
+#### 0.5.8 - 12.11.2018
+
+* npm packages updated
+* responsive fix
+* logo updated
+* logout inserted
+* now it is possible to override the default bulma's style in "global-style.scss"
+* inserted various dimensions for logo image in "index.html"
+* general fix
+
 #### 0.5.7 - 10.11.2018
 
 * Inserted validation in login modal